pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ai-dev-os"
version = "1.0.0"
description = "The AI Development Operating System — a meta-framework for orchestrating Claude Code agents across complex software projects"
readme = "README.md"
license = { text = "MIT" }
authors = [
    { name = "krzemienski", email = "krzemienski@gmail.com" }
]
keywords = ["ai", "claude", "agents", "orchestration", "development", "automation"]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Application Frameworks",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
requires-python = ">=3.10"
dependencies = [
    "click>=8.1.0",
    "pyyaml>=6.0",
    "rich>=13.0.0",
    "pydantic>=2.0.0",
]

[project.scripts]
ai-dev-os = "ai_dev_os.cli:main"

[project.urls]
Homepage = "https://github.com/krzemienski/ai-dev-operating-system"
Repository = "https://github.com/krzemienski/ai-dev-operating-system"
Documentation = "https://github.com/krzemienski/ai-dev-operating-system/docs"
"Bug Tracker" = "https://github.com/krzemienski/ai-dev-operating-system/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/ai_dev_os"]

[tool.hatch.build.targets.sdist]
include = [
    "/src",
    "/docs",
    "/README.md",
    "/LICENSE",
]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.mypy]
python_version = "3.10"
strict = true
ignore_missing_imports = true
